refactor(App, BridgeMaker): 파일 경로 수정
