+++ src/BridgeMaker.js
@@ -1,4 +1,4 @@
-const { BRIDGE_CONSTANTS } = require('./GameConstants');
+const { BRIDGE_CONSTANTS } = require('./constant/GameConstants');
 
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
